requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.8
numpy==1.18.1
pandas==1.0.1
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
